Final version, spread 2 files, one for data/dstructures, one for logic
