+++ package.json
@@ -18,7 +18,11 @@
     "deploy": "node ./deploy.js"
   },
   "dependencies": {
+    "lodash.isempty": "4.4.0",
+    "lodash.isequal": "4.5.0",
+    "lodash.partialright": "4.2.1",
     "lodash.sample": "4.2.1",
+    "lodash.unzip": "3.4.0",
     "react": "16.8.6",
     "react-copy-to-clipboard": "5.0.1",
     "react-dom": "16.8.6",
